Add files via upload
In this project, we use CNN and LSTM to identify the caption of the image. As the deep 
learning techniques are growing, huge datasets and computer power are helpful to build 
models that can generate captions for an image. This is what we are going to implement in 
this Python based project where we will use deep learning techniques like CNN and RNN. 
Image caption generator is a process which involves natural language processing and 
computer vision concepts to recognize the context of an image and present it in English. In 
this survey paper, we carefully follow some of the core concepts of image captioning and its 
common approaches. We discuss Keras library, numpy and jupyter notebooks for the making 
of this project.We also discuss about flickr_dataset and CNN used for image classification.